\name{data1}
\alias{data1}
\docType{data}
\title{
Data for the U.S. production function estimation
}
\description{
US production, captal, labour in natrual logs for the year 2005
}
\usage{data("data1")}
\format{
  A data frame with 51 observations on the following 3 variables.
  \describe{
    \item{\code{lnoutput}}{natrual log of output}
    \item{\code{lnlabor}}{natrual log of labor}
    \item{\code{lncapital}}{natrual log of capital}
  }
}
\details{
The data contains 51 observations for 50 US states and Washington DC
}
\source{
Gujarati, D. 2015, Econometrics by Example, Second edition, Palgrave.
}
\references{
See Section 2.2 of Gujarari (2015)

Kim and Choi, 2019, Choosing the Level of Significance: A Decision-theoretic Approach, Abacus: a Journal of Accounting, Finance and Business Studies. Wiley. 
}
\examples{
data(data1)
}
\keyword{datasets}
